ti-demo: Fix service file to start after multi-user.target
Without this, the launch of the service cannot be predicted and it might
launch before multi-user.target, In which case the demo will appear for
a second and then loses control over the display for some unknown reason.

Add this dependency to launch the demo at almost the end of boot.

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>