seva: Add systemd service to launch seva-launcher-aarch64
- With GPU accelerated Chromium Browser integrated in SDK 9.2 release [1],
seva-launcher-aarch64 will be responsible to create a localhost page on
port 8000 so that Chromium Browser just launches the localhost page once user
clicks Seva Store on ti-apps-launcher.

- Hence, create a systemd service which on boot will run seva-launcher-aarch64
binary that will create a localhost page on port 8000. no-browser is set to true so as
to avoid a browser launch [2].

- Restart is set to on-failure so as to ensure that the localhost page is recreated
if the seva-launcher-aarch64 exits on websocket termination [3].

- Include the service file in seva-launcher recipe & install it in ${systemd_system_unitdir}
of filesystem.

[1]: https://git.ti.com/cgit/arago-project/meta-arago/commit/?h=kirkstone&id=1c7e65c70d3d0648274b9945a94dc8a3b69e5b94

[2]: https://github.com/TexasInstruments/seva/blob/main/seva-launcher/seva-launcher.go#L30

[3]: https://github.com/TexasInstruments/seva/blob/main/seva-launcher/seva-websocket.go#L34

Signed-off-by: Chirag Shilwant <c-shilwant@ti.com>