Add a new Jailhouse OOB demo recipe for jailhouse inmate image
Add a new recipe for jailhouse OOB demo. This recipe creates a new
unit which will be used for jailhouse linux inmate and assigns
a ip address to network interface. The inmate linux then can be
pinged via that ip address.

The unit depends on login service for now (a hack) as this virtual
network interface takes time to be configured fully (it comes from
a ivshmem net module which is a part of kernel image). The service
will need to be run at a later stage to ensure that network is fully
configured.

Also add the recipe to jailhouse inmate image using PACKAGE_INSTALL

Signed-off-by: Paresh Bhagat <p-bhagat@ti.com>