recipes-kernel: linux: add support for Xen
This patch adds support to enable Xen for ti-linux-staging. This
involves a kernel fragment specifying desired defconfig options as well
as two patches:

fix-rpmsg-xen.patch: This patch fixes the rpmsg_probe failure that
happens otherwise.

drivers-...-protocol.patch: This patch [1] is required for compatibility
with a patchset [2] for xen which fixes init-dom0less and hotplugging
for dom0less domains.

[1]: https://lore.kernel.org/xen-devel/20240517011516.1451087-1-xin.wang2@amd.com/
[2]: https://lore.kernel.org/all/alpine.DEB.2.22.394.2405241552240.2557291@ubuntu-linux-20-04-desktop/

Signed-off-by: Amneesh Singh <a-singh21@ti.com>