recipes-bsp: u-boot: add mechanism to boot xen
This patch provides a custom CONFIG_BOOTCOMMAND when user has enabled
xen under DISTRO_FEATURES. This bootcommand runs a script from boot
partition, which is boot.scr by default.

The script is also provided and is compiled using mkimage as a FIT
image. This script loads the Xen image, the kernel image and the FDT
and then modifies the FDT to allow suitable Xen configuration and other
features like shared memory and UIO.

Signed-off-by: Amneesh Singh <a-singh21@ti.com>