meta-ti-security: Add WIC image
This creates a wic image of 4 partitions. dm-crypt and dm-verity require
a partition each.

The crypt partition needs to be as large as the
filesystem it contains + at least 32MB of additional buffer for the
headers.

The verity partition needs about 10% of the size of the partition it is
verifying.

Signed-off-by: Aashvij Shenai <a-shenai@ti.com>