recipes-bsp: u-boot: Add patch to fix boot of AM57x HS devices
A larger fitImage was corrupting the pointer holding address of a
register which was resulting in CPU data abort on AM57x HS devices. Add
a patch which shall be be applied on top of U-Boot source to fix boot of
AM57x HS devices by hard-coding the register address instead of
retrieving it from the pointer.

Signed-off-by: Beleswar Padhi <b-padhi@ti.com>